Unasync semaphore (#7)
* Added PythonFunctionArguments

Refactoring to normalize and consolidate function introspection (prep to reduce repeating this a lot...)
Note that this can't be used for Java signatures since there may be more than one sig per function.

* Also includes improvements to sleep jitter and max queue.

Importantly, this now uses a dictionary instead of a list, meaning the signals are thread safe (and a very slight delay drastically mitigates when multiple threads fire at almost the exact same time, since multiple may take slightly different times to get to the call_queue checks).

NOTE: the semaphore is _still_ not guaranteed an order, only one-at-a-time. Use thread blocking and sequence accordingly!

* Improved PythonFunctionArguments, fixing the `*args` `**kwargs` bit

* Thread now blocks such that the same thread can re-enter without
blocking itself.

* Switching from the weakref stuff to just use Thread IDs helps - objects
were not literal references like I expected.

* Making sure lookups can't race, since we _really_ want to be sure
of purging as we go (as well as we can)

Added clue to how to use blocking for the lifetime of a thread. (grab and own)

* TODO: there is the potential for a stale thread id to get re-used and
the call queue to mistakenly think _that_ id is the next in line. This
is because the id is always increasing, so a stale thread still cached
that has had _any_ other thread call the semaphore after it will mean
that the current min(call_queue) id is larger than the cached one.

So we can detect that. I didn't _fix_ it here, though, since I honestly
don't know how to test it without hours of grinding threads.

See note at
https://github.com/CorsoSource/metatools/commit/ea0a22c544b7840ce84c3d8ea196c4808eae4bbc#commitcomment-72835770