This seems to pass tests, even on the full test server. The change is, at its core, switching from forcing `min(call_queue)` to instead point to the `None` key.
There is a potential to race here, but since everything should be
working and fighting around the same time, this is likely fine. The time
it takes to run the semaphore'd function should be longer than the
potential for racing between lines 537 and 564, and even if it does, the
race should end with the same result.