Allow extreme race conditions to still resolve
In some extreme edge cases, it's possible for the metaclass to resolve to the wrong one on start, leading to a type error that can't be resolved. The end class def for ExtraGlobal goes into a repair/recovery routine and extracts MetaExtraGlobal to use the already initialized class of ExtraGlobal.

Also includes ExtraHardReset, which resets extra hard. It does _not_ cause MEG to be regenerated, but it does force as much as possible to re-resolved.