Read and resume from tags
If `resume` is set, then read the tags and apply those values to the state machine. NOTE: this might not work intuitively! As time clicks forward, many functions (like the easing functions) depend _strictly_ on initial conditions and the variable used for time. So unless the function is reversible or doesn't react to the time domain, merely changing a variable may not drive that variables resulting `step()` value.